requirements.txt
APScheduler==3.11.0
fastapi==0.115.12
python-dotenv==1.1.0
sendgrid==6.12.2
SQLAlchemy==2.0.40
uvicorn==0.34.2
